refactor: Race 클래스를 RacingGame으로 리팩토링
- 클래스명 변경: Race -> RacingGame
- play 메서드: 시도 횟수만큼 playRound를 호출하고 결과를 출력하도록 수정
- playRound 메서드 추가: 각 라운드에서 자동차를 이동시키고 상태를 출력
- getWinners 메서드 추가: 최고 위치에 있는 자동차들(우승자들)의 이름을 반환
- getMaxPosition 메서드 추가: 현재 자동차들 중 최대 위치를 계산
- 기타: 코드 가독성과 유지보수성을 향상시키기 위한 리팩토링
