+++ src/main/java/racingcar/domain/RacingGame.java
@@ -1,43 +1,50 @@
-package racingcar;
+package racingcar.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import racingcar.domain.Car;
-
+import java.util.stream.Collectors;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Race {
+public class RacingGame {
+    private static final int MIN_RANDOM = 0;
+    private static final int MAX_RANDOM = 9;
+    private static final int MOVE_THRESHOLD = 4;
+
     private final List<Car> cars;
     private final int attempts;
 
-    public Race(List<Car> cars, int attempts){
+    public RacingGame(List<Car> cars, int attempts){
         this.cars = cars;
         this.attempts = attempts;
     }
 
-    public void start() {
+    public void play() {
         for(int i = 0; i < attempts; i++){
-            moveAndPrintRound();
+            playRound();
             System.out.println();
         }
     }
 
-    private void moveAndPrintRound() {
+    private void playRound() {
         for(Car car : cars){
-            int moveDistance = Randoms.pickNumberInRange(0, 9);
-            car.move(moveDistance);
-            System.out.println(car.getName() + " : " + "-".repeat(car.getPosition()));  // 수정된 부분: ": " -> " : "
+            int number = Randoms.pickNumberInRange(MIN_RANDOM, MAX_RANDOM);
+            car.move(number >= MOVE_THRESHOLD);
+            System.out.println(car.toString());
         }
     }
 
     public List<String> getWinners() {
-        int maxPosition = cars.stream().mapToInt(Car::getPosition).max().orElse(0);
-        List<String> winners = new ArrayList<>();
-        for (Car car : cars) {
-            if (car.getPosition() == maxPosition) {
-                winners.add(car.getName());
-            }
-        }
-        return winners;
+       int maxPosition = getMaxPosition();
+       return cars.stream()
+               .filter(car -> car.getPosition() == maxPosition)
+               .map(Car::getName)
+               .collect(Collectors.toCollection(ArrayList::new));
+    }
+
+    private int getMaxPosition() {
+        return cars.stream()
+                .mapToInt(Car::getPosition)
+                .max()
+                .orElse(0);
     }
 }
